Fix for when this module is used in an iterative environment
In the use case documented in the README.md, if there are multiple ASGs,
and this module is then called iteratively, each iteration will now
correctly have its own output.

This may also be fixing #3.

In local testing, for a new ASG being added to a pool of existing ASGs,
the recorded output files are showing the correct values, with the string
null being recorded for the new ASG (which is correct in this context).